Refactor notes.h
Make Note enum into an enum class so that its values are scoped and
strongly typed.
Add noteName() function to convert an enum value to its name. The better
way to handle this would be to use a template or an operator overload,
but because I don't expect to have to add more values to the set of
note names I'm not concerned that the function will become
unsynchronized with the enum.